pyproject.toml
[project]
name = "rock-n-stone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.4",
    "vosk>=0.3.45",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
